+++ sdk/tables/azure-data-tables/azure/data/tables/table_shared_access_signature.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------------------------------
 from typing import Union
 
-from ._shared.models import AccountSasPermissions
+from ._models import AccountSasPermissions
 from ._models import TableServices
 from ._shared._common_conversion import _sign_string
 from ._shared._error import _validate_not_none
